+++ test/spec/sbt_spec.rb
@@ -7,7 +7,6 @@ describe 'Sbt' do
     new_default_hatchet_runner('sbt-minimal-scala-sample').tap do |app|
       app.deploy do
         expect(app.output).to include('Running: sbt compile stage')
-        expect(app.output).to include('Priming Ivy cache')
         expect(app.output).not_to include('downloading http://repo.typesafe.com/typesafe/ivy-releases/org.scala-sbt/sbt')
         expect(app.output).not_to include('Main Scala API documentation to')
         expect(app.output).to include('[info] Done packaging.')